Added initial prototype
There is the option to be a client or a server, the server will be listening for a connection from a client who will specify an address (empty for localhost) and a port to connect to.  The server will be continuously reading from a lua file a 2D position and a speed and will using those values to lerp a triangle between the current position and the desired position. The server will then send the current triangle position after the movement calculation of that frame to the client which will set his triangle's position to the one received. In debug, the solution will launce two projects of which one is empty and pointing to the executable created by the first to mimic having an application run twice, it will also search for lua scripts and other resources in specific directories of the project. In release, the main project should be selected and be built alone, the exe will then need to be placed in the same directory as the directories for lua files and other resources, thus release is meant to be treated like a finished product (an executable with resources it can access at runtime).